Just making sure it still works
Reviving just enough to make sure it works in advance of going to
gradle summit in case I want to try anything. Test suite passes
without java. With java, there are some failures that are related to
changes in java or ant, but nothing that indicates actual
functionality problems with abuild itself.